+++ packages/database/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @allends/graphene-database
 
+## 0.1.2
+
+### Patch Changes
+
+- Bug fixes
+
 ## 0.1.1
 
 ### Patch Changes